refactor: move validation into subpackage
